// src/models/estudiante.ts
interface Estudiante {
    estudiante_id: string;
    nombre: string;
    apellido_paterno: string;
    apellido_materno: string;
    codigo_matricula: string;
    correo_institucional: string;
    anio_estudio: string;
    sexo: string;
    fecha_nacimiento: string;
    celular: string;
    clave: string;
  }

  export default Estudiante;